feat/#week4 - SAA 적용
