+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
+
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -12,19 +13,23 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ATSOPTANDROID"
-        tools:targetApi="31" >
+        tools:targetApi="31">
         <activity
-            android:name="org.sopt.at.HomeActivity"
-            android:label="@string/title_activity_home"
+            android:name=".MainActivity"
+            android:exported="false"
+            android:label="@string/title_activity_main"
             android:theme="@style/Theme.ATSOPTANDROID" />
-         <!-- <activity -->
+        <activity
+            android:name=".HomeActivity"
+            android:label="@string/title_activity_home"
+            android:theme="@style/Theme.ATSOPTANDROID" /> <!-- <activity -->
         <!-- android:name=".ui.theme.ListActivity" -->
         <!-- android:exported="false" /> -->
         <activity
             android:name="org.sopt.atsoptandroid.LoginActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.ATSOPTANDROID" >
+            android:theme="@style/Theme.ATSOPTANDROID">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
