+++ app/src/main/java/org/sopt/at/component/CategoryTab.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import org.sopt.at.ui.theme.TivingAppTheme
 import org.sopt.at.ui.theme.TivingTheme
 
 @Immutable
@@ -52,7 +53,7 @@ fun CategoryTab(modifier: Modifier = Modifier) {
 @Preview(showBackground = true, backgroundColor = 0xFF000000)
 @Composable
 fun CategoryTabPreview() {
-    TivingTheme {
+    TivingAppTheme {
         CategoryTab()
     }
 }
